Chỉnh số lượng thành decimal và UI => n2 format

+++ src/Tasin.Website/Models/ViewModels/PurchaseAgreementViewModel.cs
@@ -125,7 +125,7 @@ namespace Tasin.Website.Models.ViewModels
         /// <summary>
         /// Total Quantity from all purchase orders
         /// </summary>
-        public int TotalQuantity { get; set; }
+        public decimal TotalQuantity { get; set; }
 
         /// <summary>
         /// Unit ID
